refactor: Simplify mail code
